Move stale filtering out of command package
